+++ www/grpc/validator_test.go
@@ -57,7 +57,7 @@ func TestGetValidatorByNumber(t *testing.T) {
 
 	t.Run("should return Not Found", func(t *testing.T) {
 		res, err := client.GetValidatorByNumber(tCtx, &zarb.ValidatorByNumberRequest{
-			Number: int32(val1.Number() + 1),
+			Number: val1.Number() + 1,
 		})
 		assert.Error(t, err)
 		assert.Nil(t, res)
@@ -65,12 +65,12 @@ func TestGetValidatorByNumber(t *testing.T) {
 
 	t.Run("Should return validator matching with public key and number", func(t *testing.T) {
 		res, err := client.GetValidatorByNumber(tCtx, &zarb.ValidatorByNumberRequest{
-			Number: int32(val1.Number()),
+			Number: val1.Number(),
 		})
 		assert.NoError(t, err)
 		assert.NotNil(t, res)
 		assert.Equal(t, val1.PublicKey().Bytes(), res.GetValidator().PublicKey)
-		assert.Equal(t, int32(val1.Number()), res.GetValidator().GetNumber())
+		assert.Equal(t, val1.Number(), res.GetValidator().GetNumber())
 
 	})
